update: exit if both net and net-next are up to date
The -net base is updated after net-next to split the update in two and
ease conflicts resolution.

If net-next is up to date, -net should be as well if we sync with a
common base of -net and net-next except if we force the sync to the
latest version of -net. In this case, it is still possible to quit early
but then it is required to check if -net has been updated as well.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>